+++ src/variety/heyawake.js
@@ -71,7 +71,6 @@ Graphic:{
 	bcolor_type : "GREEN",
 
 	bgcellcolor_func : "qsub1",
-	numbercolor_func : "qnum",
 
 	bbcolor : "rgb(160, 255, 191)",