+++ src/variety/lookair.js
@@ -51,7 +51,6 @@ AreaShadeGraph:{
 // 画像表示系
 Graphic:{
 	bgcellcolor_func : "qsub1",
-	numbercolor_func : "qnum",
 
 	bcolor_type : "GREEN",