+++ src/variety/yajikazu.js
@@ -66,7 +66,6 @@ Graphic:{
 	bcolor_type : "GREEN",
 
 	bgcellcolor_func : "qsub1",
-	numbercolor_func : "qnum",
 
 	fontShadecolor : "rgb(96,96,96)",